requirements.txt
jsonschema
flask
flask-cors
requests
python-dateutil

pytest
jsonschema
sphinx
sphinx-rtd-theme
